refactor: change file/dir name
