+++ src/ts/view/videoPageTemplates/index.ts
